requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
weasyprint
aiohttp
flask
python-telegram-bot==13.15
reportlab